Mushed Berries
Merging Combat Build

+++ Project/Assets/Scripts/boss.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public class boss : MonoBehaviour
 {
-    public float health = 100f;
+    public double health = 100f;
     private GameObject player;
     // Start is called before the first frame update
     void Start()
     {
         player = GameObject.FindWithTag("Player");
     }
-    public void losehealth(float amnt)
+    public void losehealth(double amnt)
     {
         health -= amnt;
     }
